feat(stories-brain): inline live stats via `bench --stats`
Extract the progress renderer into scripts/progress-view.mjs (shared by
stats.mjs) and add `bench --stats`: it tees vitest output to
results/bench.log and renders the live per-task table in the foreground,
so a run shows its rate + ETA in one terminal instead of requiring a
second one. Redraws in place on a TTY (alternate-screen), appends frames
otherwise; the child's exit drives the final frame + log-path notice.
